feat(patternplate): allow configuring patterns via single file
Allows configuring patterns by defining a single file at the root of the repository. `.commitlint-patterns.json` contains an array of patterns. Fallback to the previous method (which to be honest I do not un
derstand at all) is provided when there's no such file, or there's an error parsing the JSON.