feat(load): allow specifying helpUrl via config (#2180)
It can make configuration a bit more DRY if one can use the commitlint
configuration object to specify a custom help URL vs. having wrapper
scripts and similar things to make every invocation of `commitlint` run
with `--help-url` specified.

To allow this, `helpUrl` is added as an optional base configuration
property, and falls back to the original default URL if neither the CLI
flag nor config option is specified. The CLI option takes precedence
over the config object, if both are supplied.