fix!: fix signed-off-by detection
The signed-off-by rule does not correctly detect `Signed-off-by:`
trailers if they are not the final non-comment line of the commit. This
change ensures that sign-off trailers are correctly detected according
to the formatting rules Git uses to insert them.

The change replaces the manual parsing of the commit message with an
invocation of `git interpret-trailers`, which parses the trailers of
the commit message passed via stdin.

One known issue with this approach is that `BREAKING CHANGE` - if it
occurs grouped alongside the sign-off trailers - will break detection.
This appears to be due to the fact that Git does not accept trailers
with spaces, instead treating them as part of the body (this seems to
be an issue of the Conventional Commits specification more than
anything). This does not occur with `BREAKING-CHANGE`, which is
considered a valid trailer.

BREAKING CHANGE: `BREAKING CHANGE` in the footers block will break `Signed-off-by:` detection.

Signed-off-by: Chris Kay <chris.kay@arm.com>