feat(rules): add `trailer-exists` rule
This new rule behaves similarly to the existing `signed-off-by` rule,
but introduces what would otherwise be breaking changes to the parser in
order to resolve some issues.

Rather than attempting to parse the trailers manually, this rule will
use Git's `interpret-trailers` subcommand to parse trailers according to
Git's rules. This allows us to properly detect all valid trailers, and
does not require that the trailer to search for be the last line of the
commit message (as is required by the `signed-off-by` rule).

One downside to this approach is that Git will not detect trailers if
they are grouped alongside other trailers with whitespace in the token
(e.g. `BREAKING CHANGE`). Projects that wish to use this rule may use
`BREAKING-CHANGE`, which is synonymous with `BREAKING CHANGE` but is
considered a valid trailer.