fix(cli): remove hard coded comment char with linting `COMMIT_EDIT_MSG`
When running the cli and passing the `--edit` flag the comment char
was hard coded to a `#`. Even if there is a `commentChar` set in the
`parserOpts` the cli will override it with a `#`.

Now the cli will only set it to a `#` if its not already set in the
`parserOpts`

Fixes Issue: #2351