fix(is-ignored): ignore merge tag commit messages
Merging a tag in another branch can lead to conflicts.
In this case, the default Git commit message after resolving the conflicts will be like:
   `Merge tag 'x.y.z'`

Currently, this commit message format is being considered as an error.
So I updated the wildcards to get a consistent behavior between branch and tag merge:
 - `Merge branch 'a branch'`
 - `Merge branch 'a tag'`