fix: stop truncating the body in presence of dashes
In the current version of commitlint the body is truncated
unexpectedly when there is a pattern like
'- Something between dashes -' in the commit message body.
The reason for this behavior is that in the commit parser
package the commitlint uses (conventional-commit-parser[1]),
this pattern '- something -' is a special pattern for any
arbitrary filed. For example, if you put '-myNote-' in the
commit message body, everything that comes after this field
will be saved in an arbitrary field called myNote (Or
whatever is written between dashes) and you can use it like
other fields (header, body, etc.), but I believe we should
disable this functionality because, in the commit messages
like the one in the bug report that this commit fixes, the
user might put stack trace in the commit message body and
this way it will be truncated unexpectedly.

Fixes https://github.com/conventional-changelog/commitlint/issues/3428

[1] https://github.com/conventional-changelog/conventional-changelog/blob/master/packages/conventional-commits-parser/README.md