CI: add an Ubuntu LTS job
This way we will make sure that any dependency upgrades
don't make commitlint lose compatibility with the default
version shipped with an LTS version of Ubuntu, starting
with 24.04.

This is, IMO, important, because AFAIU commitlint lost,
some months ago, compatibility with the NodeJS version
shipped with Ubuntu 22.04, and thus it requires its users
to install NodeJS from external sources, which is a
potential security risk (and stability risk because of
needing to rely on more bleeding edge versions).