fix(git): find the closer .git
The previous method preferred finding `.git` files over `.git/` directories.
That doesn't work if, for some reason, the user a `.git/` directory inside
the repo with commitlint and that repo path is nested inside another git repo.
Ignoring the fact that this isn't a standard use of git, it still works, and
can be trivially supported by comparing the path lengths.

Priority: Low
Tests: None, but there were none in @commitlint before this, either.
Risk: Low